Initial commit for Media Analytics Software Stack
This project provides samples to demonstrate the use of Intel GPU
in simplified real-life scenarios involving media analytics.
It leverages a Software Stack which consists of the following
ingredients:
* Intel OpenVINO
* Intel Media SDK & Intel oneVPL
* Intel Media Driver
* dl-streamer (OpenVINO gstreamer plugins)
* gstreamer media plugins

Provided samples focus on the key aspects of the proper Intel software
setup and integration with other popular tools you will likely use in
your final product. We try to do our best to provide a configuration
which will demonstrate the best performance for Intel GPU
media stack.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>